+++ perum-mobile/src/query/AccountQueryProvider.tsx
@@ -8,6 +8,7 @@ import { MessagesProvider } from '../messages/MessagesProvider';
 import { RealtimeProvider } from '../realtime/RealtimeProvider';
 import { SupportProvider } from '../support/SupportProvider';
 import { HomeworkProvider } from '../homework/HomeworkProvider';
+import { AdminActionProvider } from '../support/AdminActionProvider';
 
 onlineManager.setEventListener((setOnline) => NetInfo.addEventListener((state) => setOnline(state.isConnected !== false)));
 
@@ -41,5 +42,5 @@ export function AccountQueryProvider({ children }: PropsWithChildren) {
 
   if (!account) return children;
   if (!active || active.accountId !== account.id) return null;
-  return <QueryClientProvider client={active.client}><PreferencesProvider><MessagesProvider><SupportProvider><HomeworkProvider><RealtimeProvider>{children}</RealtimeProvider></HomeworkProvider></SupportProvider></MessagesProvider></PreferencesProvider></QueryClientProvider>;
+  return <QueryClientProvider client={active.client}><PreferencesProvider><MessagesProvider><SupportProvider><AdminActionProvider><HomeworkProvider><RealtimeProvider>{children}</RealtimeProvider></HomeworkProvider></AdminActionProvider></SupportProvider></MessagesProvider></PreferencesProvider></QueryClientProvider>;
 }